refactor: :recycle: refactor files, bring to "rest" package, reformat code
