+++ backend/src/main/java/org/acme/domain/Student.java
@@ -1,19 +1,13 @@
 package org.acme.domain;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import io.quarkus.hibernate.orm.panache.PanacheEntity;
+import jakarta.persistence.*;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-import io.quarkus.hibernate.orm.panache.PanacheEntity;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
 
 /**
  * Represents a student.
@@ -21,16 +15,16 @@ import jakarta.persistence.ManyToMany;
  * @author Jet Edge
  */
 @Entity
-public class Student extends PanacheEntity{
+public class Student extends PanacheEntity {
 
     public String name;
 
     @JsonIgnoreProperties("students")
     @ManyToMany(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     @JoinTable(
-        name = "student_unit",
-        joinColumns = @JoinColumn(name = "student_id"),
-        inverseJoinColumns = @JoinColumn(name = "unit_id")
+            name = "student_unit",
+            joinColumns = @JoinColumn(name = "student_id"),
+            inverseJoinColumns = @JoinColumn(name = "unit_id")
     )
     @JsonIgnore
     public List<Unit> units = new ArrayList<Unit>();
