+++ backend/src/main/java/org/acme/rest/StudentResource.java
@@ -1,15 +1,12 @@
-package org.acme.domain;
-
-import java.util.List;
+package org.acme.rest;
 
 import jakarta.transaction.Transactional;
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import org.acme.domain.Student;
+
+import java.util.List;
 
 @Path("/students")
 public class StudentResource {
