+++ backend/src/main/java/org/acme/rest/UnitResource.java
@@ -1,16 +1,13 @@
-package org.acme.domain;
-
-import java.util.List;
+package org.acme.rest;
 
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.transaction.Transactional;
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import org.acme.domain.Unit;
+
+import java.util.List;
 
 @Path("/units")
 @RolesAllowed({"user"})
